Add new types library
