+++ tsconfig.base.json
@@ -15,6 +15,7 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
+      "@rufferal/types": ["types/src/index.ts"],
       "@rufferal/assets/*": ["assets/*"],
       "@rufferal/store": ["store/src/index.ts"],
       "@rufferal/tailwind": ["tailwind/src/index.ts"],
