+++ types/.babelrc.js
@@ -1,0 +1,21 @@
+module.exports = function (api) {
+  api.cache(true);
+
+  return {
+    presets: [
+      [
+        '@nx/react/babel',
+        {
+          runtime: 'automatic',
+          useBuiltIns: 'usage',
+        },
+      ],
+    ],
+    plugins: [],
+    env: {
+      test: {
+        presets: ['babel-preset-expo'],
+      },
+    },
+  };
+};
