+++ types/project.json
@@ -1,0 +1,9 @@
+{
+  "name": "types",
+  "$schema": "../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "types/src",
+  "projectType": "library",
+  "tags": [],
+  "// targets": "to see all targets run: nx show project types --web",
+  "targets": {}
+}
