+++ types/src/ui/pages.ts
