+++ ui/src/components/atoms/field-helper/field-helper.tsx
@@ -1,15 +1,11 @@
 import { ruffwind } from '@rufferal/tailwind';
+import { FullTextProps } from '@rufferal/types';
 import { Text } from 'react-native';
-
-export interface FieldHelperProps {
-  align?: string;
-  text?: string;
-}
 
 export const FieldHelper = ({
   align = 'text-right',
   text = 'Required',
-}: FieldHelperProps) => {
+}: FullTextProps) => {
   return (
     <Text
       style={ruffwind.style(`font-body text-2.5 leading-3 text-red-600`, align)}
