+++ ui/src/components/atoms/h2/h2.tsx
@@ -1,12 +1,9 @@
 import { ruffwind } from '@rufferal/tailwind';
+import { FullTextProps } from '@rufferal/types';
 import { Text } from 'react-native';
 
-export interface H2Props {
-  align?: string;
-  text: string | React.ReactElement;
-}
 
-export const H2 = ({ align = 'text-center', text }: H2Props) => {
+export const H2 = ({ align = 'text-center', text }: FullTextProps) => {
   return (
     <Text
       style={ruffwind.style(
