+++ ui/src/components/atoms/index.ts
@@ -12,17 +12,6 @@
  *
  */
 
-export type { AccordionSection } from './accordian/accordian';
-export type { FieldInputAreaProps } from './field-input-area/field-input-area';
-export type { FieldInputProps } from './field-input/field-input';
-export type { FieldSelectProps } from './field-select/field-select';
-export type {
-  ButtonType,
-  FieldOption,
-  FieldSize,
-  FieldState,
-} from './types/field-types';
-
 export { Accordian } from './accordian/accordian';
 export { Bottomsheet } from './bottomsheet/bottomsheet';
 export { Button } from './button/button';
