+++ ui/src/components/atoms/toggle/toggle.tsx
@@ -1,11 +1,7 @@
+import { ToggleProps } from '@rufferal/types';
 import { moderateScale } from '@rufferal/utils';
 import { useCallback } from 'react';
 import SwitchToggle from 'react-native-switch-toggle';
-
-interface ToggleProps {
-  toggleState: boolean;
-  setToggleState: () => void;
-}
 
 // 🚨🚨🚨 WARNING: SwitchToggle: Support for defaultProps will be removed from function components in a future major release. Use JavaScript default parameters instead. 🚨🚨🚨
 // BLARG: TODO: Create custom Toggle (that can support "On"/"Off" text and replace this library here and in SwitchToggle atoms)
